Moved logic to appropriate service and controller.
