+++ src/main/java/rs/teslaris/core/controller/ProceedingsPublicationController.java
@@ -3,6 +3,8 @@ package rs.teslaris.core.controller;
 import java.util.List;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,6 +20,7 @@ import rs.teslaris.core.annotation.Idempotent;
 import rs.teslaris.core.annotation.PublicationEditCheck;
 import rs.teslaris.core.dto.document.ProceedingsPublicationDTO;
 import rs.teslaris.core.dto.document.ProceedingsPublicationResponseDTO;
+import rs.teslaris.core.indexmodel.DocumentPublicationIndex;
 import rs.teslaris.core.service.interfaces.document.ProceedingsPublicationService;
 import rs.teslaris.core.service.interfaces.user.UserService;
 import rs.teslaris.core.util.jwt.JwtUtil;
@@ -46,6 +49,12 @@ public class ProceedingsPublicationController {
         return proceedingsPublicationService.findAuthorsProceedingsForEvent(eventId,
             userService.getPersonIdForUser(
                 tokenUtil.extractUserIdFromToken(bearerToken.split(" ")[1])));
+    }
+
+    @GetMapping("/event/{eventId}")
+    public Page<DocumentPublicationIndex> readAllProceedingsPublicationsForEvent(
+        @PathVariable Integer eventId, Pageable pageable) {
+        return proceedingsPublicationService.findProceedingsForEvent(eventId, pageable);
     }
 
     @PostMapping
