+++ src/main/java/rs/teslaris/core/service/interfaces/document/DocumentPublicationService.java
@@ -31,6 +31,4 @@ public interface DocumentPublicationService extends JPAService<Document> {
     Page<DocumentPublicationIndex> searchDocumentPublications(SearchRequestDTO searchRequest,
                                                               Pageable pageable,
                                                               SearchRequestType type);
-
-    Page<DocumentPublicationIndex> findProceedingsForEvent(Integer eventId, Pageable pageable);
 }
