+++ src/main/java/rs/teslaris/core/service/interfaces/document/ProceedingsPublicationService.java
@@ -1,6 +1,8 @@
 package rs.teslaris.core.service.interfaces.document;
 
 import java.util.List;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import rs.teslaris.core.dto.document.ProceedingsPublicationDTO;
 import rs.teslaris.core.dto.document.ProceedingsPublicationResponseDTO;
@@ -25,4 +27,6 @@ public interface ProceedingsPublicationService {
 
     void indexProceedingsPublication(ProceedingsPublication publication,
                                      DocumentPublicationIndex index);
+
+    Page<DocumentPublicationIndex> findProceedingsForEvent(Integer eventId, Pageable pageable);
 }
